+++ jaiva-vscode-extension/jaiva/CHANGELOG.md
@@ -7,3 +7,9 @@ All notable changes to the "jaiva" extension will be documented in this file.
 ## 1.0.0
 
 -   Initial release
+
+## 2.0.0
+
+-   Add support for `tsea` keyword
+-   Match new token types for Jaiva release 1.0.0
+-   Fix some operators finally.